requirements.txt
# Core dependencies
streamlit==1.28.2
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
openai==1.6.1
tiktoken==0.5.2
chromadb==0.4.18
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.0.3
